Feature/defi r (#1188)
* refactor and add pages

* Update env.json

* Update actions.ts

* add filtering option

* navigation fix

* add walletfee


* test

* test-signing

* add walletfee


* test

* test-signing

* Update env.json

* Update actions.ts

* add filtering option

* navigation fix

* add walletfee


* test

* test-signing

* test

* test-signing

* navigation fix

* Update actions.ts

* add filtering option

* add walletfee


* test

* test-signing

* test

* test-signing

* test

* test-signing

* test

* test-signing

---------
