+++ src/consts/index.ts
@@ -74,6 +74,7 @@ export enum WalletFilteringOptions {
   All = 'All',
   Currencies = 'Currencies',
   NFT = 'NFT',
+  SBT = 'SBT',
 }
 
 export enum SoraNetwork {
@@ -228,6 +229,7 @@ export const TranslationConsts = {
   TBC: 'TBC',
   XYK: 'XYK',
   NFT: 'NFT',
+  SBT: 'SBT',
   CEX: 'CEX',
   Polkadot: 'Polkadot',
   SORAScan: 'SORAScan',
