+++ src/index.ts
@@ -8,6 +8,7 @@ import AddAssetDetailsCard from './components/AddAsset/AddAssetDetailsCard.vue';
 import AddressBookInput from './components/AddressBook/Input.vue';
 import AssetList from './components/AssetList.vue';
 import AssetListItem from './components/AssetListItem.vue';
+import TokenLogo from './components/AssetLogos/TokenLogo.vue';
 import ConfirmDialog from './components/ConfirmDialog.vue';
 import ConnectionView from './components/Connection/ConnectionView.vue';
 import AccountConnectionList from './components/Connection/List/Account.vue';
@@ -37,9 +38,9 @@ import FormattedAddress from './components/shared/FormattedAddress.vue';
 import SyntheticSwitcher from './components/shared/SyntheticSwitcher.vue';
 import SimpleNotification from './components/SimpleNotification.vue';
 import TokenAddress from './components/TokenAddress.vue';
-import TokenLogo from './components/TokenLogo.vue';
 import TransactionHashView from './components/TransactionHashView.vue';
 import WalletBase from './components/WalletBase.vue';
+import WalletFee from './components/WalletFee.vue';
 // Shared Components
 // Other
 import * as WALLET_CONSTS from './consts';
@@ -169,6 +170,7 @@ const components = {
   WalletAccount,
   WalletAvatar,
   WalletBase,
+  WalletFee,
   AccountCard,
   AccountConfirmationOption,
   AddressBookInput,
